pyproject.toml
[project]
name = "flyte_hydra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "pydantic>=2.9.2",
    "union>=0.1.140",
    "flytekit>=1.14.0",
]
